refactore - the api folder has been moved inside the application
